feat(tasks): moving tasksService into it's own folder and file
