+++ index.js
@@ -2,7 +2,8 @@ const express = require("express");
 const bodyParser = require("body-parser");
 const cors = require("cors");
 const helmet = require("helmet");
-const { randomUUID } = require("crypto");
+
+const { tasksService } = require("./tasks/service");
 
 const app = express();
 const PORT = process.env.PORT || 8000;
@@ -12,51 +13,6 @@ app.use(cors());
 app.use(helmet());
 
 app.disable("x-powered-by");
-
-let tasks = [];
-
-const tasksService = {
-  listTasks: () => {
-    return tasks;
-  },
-  createTask: ({ title, description }) => {
-    const newTask = {
-      id: randomUUID(),
-      title,
-      description,
-    };
-
-    tasks.push(newTask);
-
-    return newTask;
-  },
-  deleteTask: (id) => {
-    const filteredTasks = tasks.filter((task) => task.id !== id);
-
-    tasks = filteredTasks;
-
-    return tasks;
-  },
-  updateTask: (id, newFields) => {
-    const newTasks = tasks.map((task) => {
-      const isTaskToBeUpdated = task.id === id;
-      if (isTaskToBeUpdated) {
-        const updatedTask = {
-          ...task,
-          ...newFields,
-        };
-
-        return updatedTask;
-      }
-
-      return task;
-    });
-
-    tasks = newTasks;
-
-    return tasks;
-  },
-};
 
 app.get("/tasks", (req, res) => {
   const tasks = tasksService.listTasks();
