+++ tasks/service.js
@@ -1,0 +1,48 @@
+const { randomUUID } = require("crypto");
+
+let tasks = [];
+
+const tasksService = {
+  listTasks: () => {
+    return tasks;
+  },
+  createTask: ({ title, description }) => {
+    const newTask = {
+      id: randomUUID(),
+      title,
+      description,
+    };
+
+    tasks.push(newTask);
+
+    return newTask;
+  },
+  deleteTask: (id) => {
+    const filteredTasks = tasks.filter((task) => task.id !== id);
+
+    tasks = filteredTasks;
+
+    return tasks;
+  },
+  updateTask: (id, newFields) => {
+    const newTasks = tasks.map((task) => {
+      const isTaskToBeUpdated = task.id === id;
+      if (isTaskToBeUpdated) {
+        const updatedTask = {
+          ...task,
+          ...newFields,
+        };
+
+        return updatedTask;
+      }
+
+      return task;
+    });
+
+    tasks = newTasks;
+
+    return tasks;
+  },
+};
+
+exports.tasksService = tasksService;
